+++ NEWS.md
@@ -1,6 +1,9 @@
 1.18.1 - 2014-mm-dd
 -------------------
 
+Enhancements:
+ * Don't loop twice over svg rows
+
 
 1.18.0 - 2014-10-14
 -------------------