+++ NEWS.md
@@ -3,6 +3,7 @@
 
 Enhancements:
  * Don't loop twice over svg rows
+ * Improve statement timeout error messages
 
 
 1.18.0 - 2014-10-14