+++ NEWS.md
@@ -1,7 +1,12 @@
-1.18.1 - 2014-mm-dd
+1.19.0 - 2014-mm-dd
 -------------------
 
+New features:
+
+ * Add more fields to error responses with hint, detail and context for SQL errors.
+
 Enhancements:
+
  * Don't loop twice over svg rows
  * Improve statement timeout error messages
  * Improve topojson output by streaming json